pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "framekit"
version = "0.5.2"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "opencv-python",
    "PyOpenGL>=3.1.7",
    "PyOpenGL_accelerate>=3.1.7",
    "Pillow",
    "tqdm",
    "pygame",
    "mutagen",
    "librosa",
]

[tool.setuptools]
packages = ["framekit"]